+++ CHANGELOG.md
@@ -127,6 +127,7 @@
 * [X] breeze-serialization 提供缺失的基础的基于kotlin-jvm的KSerializer。
 * [X] breeze-dsl MarkdownDsl 支持标题的特性（一般是id）。 
 * [X] 精简注解。
+* [X] breeze-core 参照标准库，重命名CollectionExtensions中的部分方法。
 * [ ] breeze-dsl 实现`YamlDsl`。（不推荐使用，但是作为规范）
 * [ ] breeze-serialization 提供yaml的序列化实现。